move photo processing to settings
