+++ app/(tabs)/stats.tsx
@@ -1,12 +1,4 @@
-import {
-  View,
-  StyleSheet,
-  ScrollView,
-  Text,
-  Button,
-  SafeAreaView,
-  Alert,
-} from "react-native";
+import { View, StyleSheet, ScrollView, SafeAreaView } from "react-native";
 import {
   MediaStats,
   initialMediaStats,
@@ -16,16 +8,13 @@ import { useState, useEffect } from "react";
 import { useIsFocused } from "@react-navigation/native";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import StatsCarousel from "@/components/StatCarousel";
-import { useMediaProcessing } from "@/hooks/useMediaProcessing";
 import StatGraph from "@/components/StatGraph";
 import StatPieChart from "@/components/StatPieChart";
-import GradientButton from "@/components/GradientButton";
 import GradientText from "@/components/GradientText";
 
 export default function StatsScreen() {
   const [stats, setStats] = useState<MediaStats>(initialMediaStats);
   const isFocused = useIsFocused();
-  const { status, startProcessing } = useMediaProcessing();
 
   const loadStoredStats = async () => {
     try {
@@ -43,16 +32,6 @@ export default function StatsScreen() {
       loadStoredStats();
     }
   }, [isFocused]);
-
-  const reProcess = async () => {
-    const permissionGranted = await startProcessing();
-    if (!permissionGranted) {
-      Alert.alert(
-        "Please grant media library permissions to analyze your photos",
-      );
-      return;
-    }
-  };
 
   return (
     <SafeAreaView style={{ flex: 1, backgroundColor: "#212121" }}>
@@ -112,39 +91,7 @@ export default function StatsScreen() {
           <StatPieChart data={stats.fileTypes} title="File Type" />
           <StatPieChart data={stats.timeOfDay} title="Time Of Day" />
         </View>
-        <View style={{ alignItems: "center" }}>
-          <GradientButton
-            text="Process Library"
-            onPress={reProcess}
-            style={buttonStyle}
-          />
-        </View>
-        {status.isProcessing && (
-          <View style={{ alignItems: "center", padding: 12 }}>
-            <Text style={{ color: "white" }}>Progress: {status.progress}</Text>
-          </View>
-        )}
       </ScrollView>
     </SafeAreaView>
   );
 }
-
-const buttonStyle = StyleSheet.create({
-  buttonGradient: {
-    width: 325,
-    height: 75,
-    borderRadius: 22.5,
-  },
-  buttonText: {
-    color: "#FFFFFF",
-    fontSize: 26,
-    fontWeight: "bold",
-  },
-  buttonContainer: {
-    flex: 1,
-    width: 325,
-    height: 75,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
